Miscellaneous tidying.
