+++ config/BTS_2HstrCURRENT_4KickerAMP_4KickerTIME_MaxIE_MinPPX
@@ -2,14 +2,14 @@
 S'results'
 p1
 (lp2
-(i__main__
-mr_representation
+(idlsoo.config
+MrRepresentation
 p3
 (dp4
 S'mr_obj'
 p5
-(idls_optimiser_util
-dls_measurement_var
+(idlsoo.config
+DlsMeasurementVar
 p6
 (dp7
 S'delay'
@@ -45,13 +45,13 @@ sS'ar_label'
 p21
 S'CS-DI-XFER-01:BS-MS'
 p22
-sba(i__main__
-mr_representation
+sba(idlsoo.config
+MrRepresentation
 p23
 (dp24
 g5
-(idls_optimiser_util
-dls_measurement_var
+(idlsoo.config
+DlsMeasurementVar
 p25
 (dp26
 g8
@@ -81,8 +81,8 @@ p32
 sbasS'parameters'
 p33
 (lp34
-(i__main__
-mp_group_representation
+(idlsoo.config
+MpGroupRepresentation
 p35
 (dp36
 S'relative_setting'
@@ -95,11 +95,11 @@ sS'ap_label'
 p39
 S'SR01S-PC-SIK-01:SETI'
 p40
-sS'mp_representations'
+sS'MpRepresentations'
 p41
 (lp42
-(i__main__
-mp_representation
+(idlsoo.config
+MpRepresentation
 p43
 (dp44
 g19
@@ -110,8 +110,8 @@ p46
 g40
 sS'mp_obj'
 p47
-(idls_optimiser_util
-dls_param_var
+(idlsoo.config
+DlsParamVar
 p48
 (dp49
 g8
@@ -125,8 +125,8 @@ p51
 F-20.0
 sg19
 g45
-sba(i__main__
-mp_group_representation
+sba(idlsoo.config
+MpGroupRepresentation
 p52
 (dp53
 g37
@@ -138,8 +138,8 @@ S'SR01S-PC-SIK-02:SETI'
 p54
 sg41
 (lp55
-(i__main__
-mp_representation
+(idlsoo.config
+MpRepresentation
 p56
 (dp57
 g19
@@ -148,8 +148,8 @@ p58
 sg46
 g54
 sg47
-(idls_optimiser_util
-dls_param_var
+(idlsoo.config
+DlsParamVar
 p59
 (dp60
 g8
@@ -161,8 +161,8 @@ sbsbasg51
 F-20.0
 sg19
 g58
-sba(i__main__
-mp_group_representation
+sba(idlsoo.config
+MpGroupRepresentation
 p61
 (dp62
 g37
@@ -174,8 +174,8 @@ S'SR01S-PC-SIK-03:SETI'
 p63
 sg41
 (lp64
-(i__main__
-mp_representation
+(idlsoo.config
+MpRepresentation
 p65
 (dp66
 g19
@@ -184,8 +184,8 @@ p67
 sg46
 g63
 sg47
-(idls_optimiser_util
-dls_param_var
+(idlsoo.config
+DlsParamVar
 p68
 (dp69
 g8
@@ -197,8 +197,8 @@ sbsbasg51
 F-20.0
 sg19
 g67
-sba(i__main__
-mp_group_representation
+sba(idlsoo.config
+MpGroupRepresentation
 p70
 (dp71
 g37
@@ -210,8 +210,8 @@ S'SR01S-PC-SIK-04:SETI'
 p72
 sg41
 (lp73
-(i__main__
-mp_representation
+(idlsoo.config
+MpRepresentation
 p74
 (dp75
 g19
@@ -220,8 +220,8 @@ p76
 sg46
 g72
 sg47
-(idls_optimiser_util
-dls_param_var
+(idlsoo.config
+DlsParamVar
 p77
 (dp78
 g8
@@ -233,8 +233,8 @@ sbsbasg51
 F-20.0
 sg19
 g76
-sba(i__main__
-mp_group_representation
+sba(idlsoo.config
+MpGroupRepresentation
 p79
 (dp80
 g37
@@ -246,8 +246,8 @@ S'SR01S-PC-EVR-01:TI:SIK-01'
 p81
 sg41
 (lp82
-(i__main__
-mp_representation
+(idlsoo.config
+MpRepresentation
 p83
 (dp84
 g19
@@ -256,8 +256,8 @@ p85
 sg46
 g81
 sg47
-(idls_optimiser_util
-dls_param_var
+(idlsoo.config
+DlsParamVar
 p86
 (dp87
 g8
@@ -269,8 +269,8 @@ sbsbasg51
 F-0.024
 sg19
 g85
-sba(i__main__
-mp_group_representation
+sba(idlsoo.config
+MpGroupRepresentation
 p88
 (dp89
 g37
@@ -282,8 +282,8 @@ S'SR01S-PC-EVR-01:TI:SIK-02'
 p90
 sg41
 (lp91
-(i__main__
-mp_representation
+(idlsoo.config
+MpRepresentation
 p92
 (dp93
 g19
@@ -292,8 +292,8 @@ p94
 sg46
 g90
 sg47
-(idls_optimiser_util
-dls_param_var
+(idlsoo.config
+DlsParamVar
 p95
 (dp96
 g8
@@ -305,8 +305,8 @@ sbsbasg51
 F-0.024
 sg19
 g94
-sba(i__main__
-mp_group_representation
+sba(idlsoo.config
+MpGroupRepresentation
 p97
 (dp98
 g37
@@ -318,8 +318,8 @@ S'SR01S-PC-EVR-01:TI:SIK-03'
 p99
 sg41
 (lp100
-(i__main__
-mp_representation
+(idlsoo.config
+MpRepresentation
 p101
 (dp102
 g19
@@ -328,8 +328,8 @@ p103
 sg46
 g99
 sg47
-(idls_optimiser_util
-dls_param_var
+(idlsoo.config
+DlsParamVar
 p104
 (dp105
 g8
@@ -341,8 +341,8 @@ sbsbasg51
 F-0.024
 sg19
 g103
-sba(i__main__
-mp_group_representation
+sba(idlsoo.config
+MpGroupRepresentation
 p106
 (dp107
 g37
@@ -354,8 +354,8 @@ S'SR01S-PC-EVR-01:TI:SIK-04'
 p108
 sg41
 (lp109
-(i__main__
-mp_representation
+(idlsoo.config
+MpRepresentation
 p110
 (dp111
 g19
@@ -364,8 +364,8 @@ p112
 sg46
 g108
 sg47
-(idls_optimiser_util
-dls_param_var
+(idlsoo.config
+DlsParamVar
 p113
 (dp114
 g8
@@ -377,8 +377,8 @@ sbsbasg51
 F-0.024
 sg19
 g112
-sba(i__main__
-mp_group_representation
+sba(idlsoo.config
+MpGroupRepresentation
 p115
 (dp116
 g37
@@ -390,8 +390,8 @@ S'BS-PC-HSTR-06:SETI'
 p117
 sg41
 (lp118
-(i__main__
-mp_representation
+(idlsoo.config
+MpRepresentation
 p119
 (dp120
 g19
@@ -400,8 +400,8 @@ p121
 sg46
 g117
 sg47
-(idls_optimiser_util
-dls_param_var
+(idlsoo.config
+DlsParamVar
 p122
 (dp123
 g8
@@ -413,8 +413,8 @@ sbsbasg51
 F-1.0
 sg19
 g121
-sba(i__main__
-mp_group_representation
+sba(idlsoo.config
+MpGroupRepresentation
 p124
 (dp125
 g37
@@ -426,8 +426,8 @@ S'BS-PC-HSTR-07:SETI'
 p126
 sg41
 (lp127
-(i__main__
-mp_representation
+(idlsoo.config
+MpRepresentation
 p128
 (dp129
 g19
@@ -436,8 +436,8 @@ p130
 sg46
 g126
 sg47
-(idls_optimiser_util
-dls_param_var
+(idlsoo.config
+DlsParamVar
 p131
 (dp132
 g8
