+++ dlsoo/util.py
@@ -1,28 +1,45 @@
-'''
-Version 2
-2017-07-30
-
-This version of util is used in basic optimisation. The important class to read through are sim_machine_interactor_bulk_base and
-dls_machine_interactor_bulk_base. This contains the methods for starting and stopping injection. The measure_results function is
-also an important part to understand; this is how the interactor deals with noisy data.
-
-IMPORTANT KEY:
-        ARs: Algorithm results (objectives)
-        APs: Algorithm parameters
-        MRs: Machine results (objectives)
-        MPs: Machine parameters
-
-'''
 from __future__ import division
 
 import time
 import math
 import pickle
-import model
 import ca_abstraction_mapping
 
 from cothread.catools import caget, caput
 import cothread
+
+
+def extract_column(matrix, colnum):
+    col = []
+    for i in matrix:
+        col.append(i[colnum])
+    return col
+
+
+def mean(x):
+    return sum(x)/len(x)
+
+
+def extract_numbers(string1):
+    #takes in a list and extracts the numbers with : ; around them to then stor in a list.
+    collect = False
+    collector = ''
+    numbers = []
+    for i in string1:
+        if i == ':':
+            collect = True
+        elif i == ';':
+            collect = False
+            numbers.append(float(collector))
+            collector = ''
+        elif collect:
+            collector += i
+    return numbers
+
+
+def save_object(obj, filename):
+    with open(filename, 'wb') as output:
+        pickle.dump(obj, output)
 
 
 def update_beam_current_bounds(I_min, I_max):
@@ -34,6 +51,7 @@ def update_beam_current_bounds(I_min, I_max):
      beam_current_bounds = [I_min, I_max]
      print 'BEAM CURRENT BOUNDS ARE ',beam_current_bounds
 
+
 def abstract_caget(pv):
     """
     standard channel access 'get' function using cothread
@@ -43,7 +61,6 @@ def abstract_caget(pv):
     else:
         return caget(pv)
 
-#-------------------------------------------------------------------
 
 def abstract_caput(pv, value):
     """
@@ -51,7 +68,6 @@ def abstract_caput(pv, value):
     """
     caput(pv, value)
 
-#-------------------------------------------------------------------
 
 def set_params(param_vars, settings, set_command):
     """
@@ -71,7 +87,6 @@ def set_params(param_vars, settings, set_command):
     # Sleep for the appropriate amount of time
     cothread.Sleep(max_delay)
 
-#-------------------------------------------------------------------
 
 def measure_results(measurement_vars, get_command):
     """
@@ -132,7 +147,6 @@ def measure_results(measurement_vars, get_command):
 
     return results
 
-#-------------------------------------------------------------------
 
 def find_group_a_bounds(param_var_min, param_var_max, initial_values, set_relative):
     '''
@@ -187,7 +201,6 @@ def find_group_a_bounds(param_var_min, param_var_max, initial_values, set_relati
     print (min, max)
     return (min, max)
 
-#-------------------------------------------------------------------------
 
 def save_details_file(object):
     """
@@ -224,634 +237,6 @@ def save_details_file(object):
 #-----------------------------------------PARAMETER AND OBJECTIVE PYTHON OBJECTS USED IN MAIN.PY------------------------------------#
 
 #these classes are used in the 'add_pv' type functions in many classes in main.py
-
-class dls_param_var:
-    def __init__(self, pv, delay):
-        self.pv = pv
-        self.delay = delay
-
-        self.initial_setting = None
-
-
-class dls_measurement_var:
-    def __init__(self, pv, min_counts, delay):
-        self.pv = pv
-        self.min_counts = min_counts
-        self.delay = delay
-
-###################################################  MAIN INTERACTOR FOR BASIC (MACHINE) #############################################
-
-#Many of the functions within this class are used in main.py and the algorithm files. The most important to understand is get_mr
-
-class dls_machine_interactor_bulk_base:
-
-    def __init__(self,
-                 param_var_groups=None,
-                 measurement_vars=None,
-                 set_relative=None,
-                 results=None):
-
-        self.param_var_groups = param_var_groups
-        self.measurement_vars = measurement_vars
-        self.results = results
-
-        self.param_vars = []
-        for group in self.param_var_groups:
-            for param in group:
-                self.param_vars.append(param)
-
-        if set_relative == None:
-            self.set_relative = []
-
-            for i in self.param_var_groups:
-                self.set_relative.append(False)
-
-        # If we need to do relative setting, we need the initial values
-        if set_relative != None:
-            self.initial_values = self.get_mp()
-            self.set_relative = set_relative
-
-
-        ''' We create a dictionary to store the input ap keys, with the output mp values '''
-        self.ap_to_mp_store = {}
-
-
-    def save_details_file(self):
-        return save_details_file(self)
-
-    def get_pv(self, pv):
-        return abstract_caget(pv)
-
-    def set_pv(self, pv, value):
-        caput(pv, value)
-
-    def ap_to_mp(self, aps):
-
-        mps = []
-
-        mpsindex = 0
-        for ngroup, group in enumerate(self.param_var_groups):
-
-            for nparam, param in enumerate(group):
-
-                if self.set_relative[ngroup] == True:
-                    mps.append(self.initial_values[mpsindex] + aps[ngroup])
-
-                else:
-                    mps.append(aps[ngroup])
-
-                mpsindex += 1
-
-        ''' Store this mapping in the ap_to_mp_store dictionary '''
-        self.ap_to_mp_store[tuple(aps)] = tuple(mps)
-        #print self.ap_to_mp_store
-
-        return mps
-
-    def mp_to_ap(self, mps):
-
-        aps = []
-
-        mpsindex = 0
-        for ngroup, group in enumerate(self.param_var_groups):
-            print mpsindex
-            if self.set_relative[ngroup]:
-                aps.append(mps[mpsindex] - self.initial_values[mpsindex])
-            elif not self.set_relative[ngroup]:
-                aps.append(mps[mpsindex])
-
-            for nparam, param in enumerate(group):
-                mpsindex += 1
-
-
-        print "mps: {0}".format(mps)
-        print "aps: {0}".format(aps)
-        print "initial values: {0}".format(self.initial_values)
-
-        return aps
-
-    def mr_to_ar(self, mrs):
-        #converts a set of machine results to algorithm results
-        ars = []
-
-        mr_to_ar_sign = [mrr.mr_to_ar_sign for mrr in self.results]
-        for mr, sign in zip(mrs, mr_to_ar_sign):
-            if sign == '+':
-                ars.append(mr)
-            elif sign == '-':
-                ars.append(-mr)
-
-        return ars
-
-    def set_mp(self, mps):
-        set_params(self.param_vars, mps, caput)
-
-    def get_mp(self):
-        mps = []
-        for param in self.param_vars:
-            print "THIS IS THE PV!: {0}".format(param.pv)
-            print type(param.pv)
-            print param.pv.encode('ascii', 'ignore')
-            print type(param.pv.encode('ascii', 'ignore'))
-            mps.append(abstract_caget(param.pv))
-
-        return mps
-
-    def get_mr(self):
-        mrs = measure_results(self.measurement_vars, abstract_caget)
-        return mrs
-
-    def set_ap(self, aps):
-        mps = self.ap_to_mp(aps)
-        self.set_mp(mps)
-
-    def get_ap(self):
-        mps = self.get_mp()
-        aps = self.mp_to_ap(mps)
-        return aps
-
-    def get_ar(self):
-        mrs = self.get_mr()
-        ars = self.mr_to_ar(mrs)
-        return ars
-
-    def find_a_bounds(self, param_var_min, param_var_max):
-
-        min_bounds = []
-        max_bounds = []
-
-        mpsindex = 0
-        for ngroup, group in enumerate(self.param_var_groups):
-
-            if self.set_relative[ngroup]:
-                min = None
-                max = None
-
-                for param in group:
-                    amount_above = param_var_max[mpsindex] - self.initial_values[mpsindex]
-                    amount_below = param_var_min[mpsindex] - self.initial_values[mpsindex]
-
-                    if min != None:
-                        if amount_below > min:
-                            min = amount_below
-                    else:
-                        min = amount_below
-
-                    if max != None:
-                        if amount_above < max:
-                            max = amount_above
-                    else:
-                        max = amount_above
-
-                    mpsindex += 1
-
-            else:
-                min = None
-                max = None
-
-                for param in group:
-                    if min != None:
-                        if param_var_min[mpsindex] > min:
-                            min = param_var_min[mpsindex]
-                    else:
-                        min = param_var_min[mpsindex]
-
-                    if max != None:
-                        if param_var_max[mpsindex] < max:
-                            max = param_var_max[mpsindex]
-                    else:
-                        max = param_var_max[mpsindex]
-
-                    mpsindex += 1
-
-
-            min_bounds.append(min)
-            max_bounds.append(max)
-
-        print (min_bounds, max_bounds)
-        return (min_bounds, max_bounds)
-
-    def string_ap_to_mp_store(self):
-        print self.ap_to_mp_store
-        return pickle.dumps(self.ap_to_mp_store)
-
-###################################################  MAIN INTERACTOR FOR BASIC (SIMULATION) #############################################
-
-#Many of the functions within this class are used in main.py and the algorithm files. The most important to understand is get_mr.
-
-class sim_machine_interactor_bulk_base:
-
-    def __init__(self,
-                 param_var_groups=None,
-                 measurement_vars=None,
-                 set_relative=None,
-                 results=None):
-
-        self.param_var_groups = param_var_groups
-        self.measurement_vars = measurement_vars
-        self.results = results
-
-        self.param_vars = []
-        for group in self.param_var_groups:
-            for param in group:
-                self.param_vars.append(param)
-
-        if set_relative == None:
-            self.set_relative = []
-
-            for i in self.param_var_groups:
-                self.set_relative.append(False)
-
-        # If we need to do relative setting, we need the initial values
-        if set_relative != None:
-            self.initial_values = self.get_mp()
-            self.set_relative = set_relative
-
-
-        ''' We create a dictionary to store the input ap keys, with the output mp values '''
-        self.ap_to_mp_store = {}
-
-
-    def save_details_file(self):
-        return save_details_file(self)
-
-    def get_pv(self, pv):
-        return model.caget(pv)
-
-    def set_pv(self, pv, value):
-        model.caput(pv, value)
-
-    def ap_to_mp(self, aps):
-
-        mps = []
-
-        mpsindex = 0
-        for ngroup, group in enumerate(self.param_var_groups):
-
-            for nparam, param in enumerate(group):
-
-                if self.set_relative[ngroup] == True:
-                    mps.append(self.initial_values[mpsindex] + aps[ngroup])
-
-                else:
-                    mps.append(aps[ngroup])
-
-                mpsindex += 1
-
-        ''' Store this mapping in the ap_to_mp_store dictionary '''
-        self.ap_to_mp_store[tuple(aps)] = tuple(mps)
-        #print self.ap_to_mp_store
-
-        return mps
-
-    def mp_to_ap(self, mps):
-
-        aps = []
-
-        mpsindex = 0
-        for ngroup, group in enumerate(self.param_var_groups):
-            print mpsindex
-            if self.set_relative[ngroup]:
-                aps.append(mps[mpsindex] - self.initial_values[mpsindex])
-            elif not self.set_relative[ngroup]:
-                aps.append(mps[mpsindex])
-
-            for nparam, param in enumerate(group):
-                mpsindex += 1
-
-
-        print "mps: {0}".format(mps)
-        print "aps: {0}".format(aps)
-        print "initial values: {0}".format(self.initial_values)
-
-        return aps
-
-    def mr_to_ar(self, mrs):
-        #converts a set of machine results to algorithm results
-        ars = []
-
-        mr_to_ar_sign = [mrr.mr_to_ar_sign for mrr in self.results]
-        for mr, sign in zip(mrs, mr_to_ar_sign):
-            if sign == '+':
-                ars.append(mr)
-            elif sign == '-':
-                ars.append(-mr)
-
-        return ars
-
-    def set_mp(self, mps):
-        set_params(self.param_vars, mps, model.caput)
-
-    def get_mp(self):
-        mps = []
-        for param in self.param_vars:
-            mps.append(model.caget(param.pv))
-
-        return mps
-
-    def get_mr(self):
-        mrs = measure_results(self.measurement_vars, model.caget)
-        return mrs
-
-    def set_ap(self, aps):
-        mps = self.ap_to_mp(aps)
-        self.set_mp(mps)
-
-    def get_ap(self):
-        mps = self.get_mp()
-        aps = self.mp_to_ap(mps)
-        return aps
-
-    def get_ar(self):
-        mrs = self.get_mr()
-        ars = self.mr_to_ar(mrs)
-        return ars
-
-    def find_a_bounds(self, param_var_min, param_var_max):
-
-        min_bounds = []
-        max_bounds = []
-
-        mpsindex = 0
-        for ngroup, group in enumerate(self.param_var_groups):
-
-            if self.set_relative[ngroup]:
-                min = None
-                max = None
-
-                for param in group:
-                    amount_above = param_var_max[mpsindex] - self.initial_values[mpsindex]
-                    amount_below = param_var_min[mpsindex] - self.initial_values[mpsindex]
-
-                    if min != None:
-                        if amount_below > min:
-                            min = amount_below
-                    else:
-                        min = amount_below
-
-                    if max != None:
-                        if amount_above < max:
-                            max = amount_above
-                    else:
-                        max = amount_above
-
-                    mpsindex += 1
-
-            else:
-                min = None
-                max = None
-
-                for param in group:
-                    if min != None:
-                        if param_var_min[mpsindex] > min:
-                            min = param_var_min[mpsindex]
-                    else:
-                        min = param_var_min[mpsindex]
-
-                    if max != None:
-                        if param_var_max[mpsindex] < max:
-                            max = param_var_max[mpsindex]
-                    else:
-                        max = param_var_max[mpsindex]
-
-                    mpsindex += 1
-
-
-            min_bounds.append(min)
-            max_bounds.append(max)
-
-        print (min_bounds, max_bounds)
-        return (min_bounds, max_bounds)
-
-    def string_ap_to_mp_store(self):
-        print self.ap_to_mp_store
-        return pickle.dumps(self.ap_to_mp_store)
-
-######################################################### END OF USEFUL CLASSES/FUNCTIONS ##########################################
-
-# the remaining classes/functions are not used in injection that often and are not important in understanding how the file works.
-
-class dls_machine_interactor_bulk_base_inj_control:
-
-    def __init__(self, param_var_groups=None, measurement_vars_noinj=None,
-                 measurement_vars_inj=None, set_relative=None):
-
-        self.param_var_groups = param_var_groups
-        self.measurement_vars_noinj = measurement_vars_noinj
-        self.measurement_vars_inj = measurement_vars_inj
-
-        self.param_vars = []
-        for group in self.param_var_groups:
-            for param in group:
-                self.param_vars.append(param)
-
-        if set_relative == None:
-            self.set_relative = []
-
-            for i in self.param_var_groups:
-                self.set_relative.append(False)
-
-        # If we need to do relative setting, we need the initial values
-        if set_relative != None:
-            self.initial_values = self.get_mp()
-            self.set_relative = set_relative
-
-        ''' We create a dictionary to store the input ap keys, with the output mp values '''
-        self.ap_to_mp_store = {}
-
-    def save_details_file(self):
-        return save_details_file(self)
-
-    def get_pv(self, pv):
-        return abstract_caget(pv)
-
-    def set_pv(self, pv, value):
-        caput(pv, value)
-
-    def ap_to_mp(self, aps):
-
-        mps = []
-
-        mpsindex = 0
-        for ngroup, group in enumerate(self.param_var_groups):
-
-            for nparam, param in enumerate(group):
-
-                if self.set_relative[ngroup] == True:
-                    mps.append(self.initial_values[mpsindex] + aps[ngroup])
-
-                else:
-                    mps.append(aps[ngroup])
-
-                mpsindex += 1
-
-        ''' Store this mapping in the ap_to_mp_store dictionary '''
-        self.ap_to_mp_store[tuple(aps)] = tuple(mps)
-
-        return mps
-
-    def mp_to_ap(self, mps):
-
-        aps = []
-
-        mpsindex = 0
-        for ngroup, group in enumerate(self.param_var_groups):
-            print mpsindex
-            if self.set_relative[ngroup]:
-                aps.append(mps[mpsindex] - self.initial_values[mpsindex])
-            elif not self.set_relative[ngroup]:
-                aps.append(mps[mpsindex])
-
-            for nparam, param in enumerate(group):
-                mpsindex += 1
-
-        print "mps: {0}".format(mps)
-        print "aps: {0}".format(aps)
-        print "initial values: {0}".format(self.initial_values)
-
-        return aps
-
-    def mr_to_ar(self, mrs):
-        return mrs
-
-    def set_mp(self, mps):
-        set_params(self.param_vars, mps, caput)
-
-    def get_mp(self):
-        mps = []
-        for param in self.param_vars:
-            mps.append(abstract_caget(param.pv))
-
-        return mps
-
-    # -------------------------- MOST IMPORTANT FUNCTION IN CLASS FOR INJECTION CONTROL ----------------------#
-
-    def get_mr(self):
-        global beam_current_bounds
-        get_command = abstract_caget
-        beam_current_max_warning = False
-
-        ''' First measure the injection results '''
-        # Begin injecting
-        print "Start injection"
-        caput('LI-TI-MTGEN-01:START', 1)
-        cothread.Sleep(0.1)
-        caput('LI-TI-MTGEN-01:START', 0)
-        cothread.Sleep(4.0)
-
-        beam_current = get_command('SR-DI-DCCT-01:SIGNAL')
-        while beam_current < beam_current_bounds[0]:
-            print 'waiting for beam current to rise above ', \
-            beam_current_bounds[0]
-            cothread.Sleep(1)
-            beam_current = get_command('SR-DI-DCCT-01:SIGNAL')
-            print '...'
-
-        run = True
-        start_time = time.time()
-
-        mrs_inj = measure_results(self.measurement_vars_inj, abstract_caget)
-
-        ''' Now for the non-injection measurements '''
-
-        if get_command('SR-DI-DCCT-01:SIGNAL') > beam_current_bounds[1]:
-            beam_current_max_warning = True
-
-        # Stop injection
-        print "Stop injection"
-        caput('LI-TI-MTGEN-01:STOP', 1)
-        cothread.Sleep(0.1)
-        caput('LI-TI-MTGEN-01:STOP', 0)
-        cothread.Sleep(1)
-
-        mrs_noinj = measure_results(self.measurement_vars_noinj,
-                                    abstract_caget)
-
-        ''' Now combine the results into a single list '''
-
-        results = mrs_noinj + mrs_inj
-
-        mrs = results
-
-        return mrs, beam_current_max_warning
-
-    # -----------------------------------------------------------------------
-
-    def set_ap(self, aps):
-        mps = self.ap_to_mp(aps)
-        self.set_mp(mps)
-
-    def get_ap(self):
-        mps = self.get_mp()
-        aps = self.mp_to_ap(mps)
-        return aps
-
-    def get_ar(self):
-        mrs, beam_current_warnings = self.get_mr()
-        ars = self.mr_to_ar(mrs)
-        return ars, beam_current_warnings
-
-    def find_a_bounds(self, param_var_min, param_var_max):
-
-        min_bounds = []
-        max_bounds = []
-
-        mpsindex = 0
-        for ngroup, group in enumerate(self.param_var_groups):
-
-            if self.set_relative[ngroup]:
-                min = None
-                max = None
-
-                for param in group:
-                    amount_above = param_var_max[mpsindex] - \
-                                   self.initial_values[mpsindex]
-                    amount_below = param_var_min[mpsindex] - \
-                                   self.initial_values[mpsindex]
-
-                    if min != None:
-                        if amount_below > min:
-                            min = amount_below
-                    else:
-                        min = amount_below
-
-                    if max != None:
-                        if amount_above < max:
-                            max = amount_above
-                    else:
-                        max = amount_above
-
-                    mpsindex += 1
-
-            else:
-                min = None
-                max = None
-
-                for param in group:
-                    if min != None:
-                        if param_var_min[mpsindex] > min:
-                            min = param_var_min[mpsindex]
-                    else:
-                        min = param_var_min[mpsindex]
-
-                    if max != None:
-                        if param_var_max[mpsindex] < max:
-                            max = param_var_max[mpsindex]
-                    else:
-                        max = param_var_max[mpsindex]
-
-                    mpsindex += 1
-
-            min_bounds.append(min)
-            max_bounds.append(max)
-
-        print (min_bounds, max_bounds)
-        return (min_bounds, max_bounds)
-
-    def string_ap_to_mp_store(self):
-        print self.ap_to_mp_store
-        return pickle.dumps(self.ap_to_mp_store)
-
 
 class measurement:
